database-2025/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.44
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2